backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
SQLAlchemy>=2.0.36
python-dotenv>=1.0.1
httpx>=0.27.2
PyMuPDF>=1.24.0
pydantic[email]>=2.9.2
pydantic-settings>=2.6.1
google-genai>=0.17.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
anyio>=4.0.0
itsdangerous>=2.1.0
